+++ rules/prettier.js
@@ -1,0 +1,13 @@
+module.exports = {
+  rules: {
+    'prettier/prettier': [
+      'error',
+      {
+        bracketSpacing: false,
+        endOfLine: 'lf',
+        printWidth: 124,
+        singleQuote: true
+      }
+    ]
+  }
+};