+++ rules/style.js
@@ -1,32 +1,17 @@
 module.exports = {
   rules: {
-    'block-spacing': 'off',
-    'brace-style': 'error',
-    'camelcase': 'error',
-    'comma-dangle': ['error', 'never'],
+    camelcase: 'error',
     'func-names': 'off',
     'func-style': ['error', 'expression'],
-    'function-paren-newline': ['error', 'never'],
-    'id-length': ['error', {
-      min: 1,
-      max: 25
-    }],
-    'keyword-spacing': 'error',
+    'id-length': [
+      'error',
+      {
+        min: 1,
+        max: 25
+      }
+    ],
     'line-comment-position': ['error', {position: 'above'}],
-    'lines-around-comment': ['error', {
-      allowArrayStart: false,
-      allowArrayEnd: false,
-      allowBlockStart: true,
-      allowBlockEnd: false,
-      allowObjectStart: false,
-      allowObjectEnd: false,
-      afterBlockComment: false,
-      afterLineComment: false,
-      beforeBlockComment: true,
-      beforeLineComment: true
-    }],
     'max-depth': ['error', 4],
-    'max-len': 'off',
     'max-lines-per-function': 'error',
     'max-nested-callbacks': ['error', 4],
     'max-params': ['error', 6],
@@ -36,12 +21,6 @@ module.exports = {
     'no-inline-comments': 'error',
     'no-negated-condition': 'error',
     'no-underscore-dangle': 'off',
-    'object-curly-newline': ['error', {
-      ObjectExpression: {consistent: true},
-      ObjectPattern: 'never'
-    }],
-    'object-curly-spacing': ['error', 'never'],
-    'padded-blocks': ['error', {classes: 'always'}],
     'padding-line-between-statements': [
       'error',
       {blankLine: 'always', prev: '*', next: 'return'},
@@ -61,22 +40,17 @@ module.exports = {
       {blankLine: 'any', prev: 'cjs-import', next: '*'},
       {blankLine: 'any', prev: 'cjs-import', next: 'cjs-import'}
     ],
-    'quote-props': ['error', 'consistent-as-needed', {keywords: true}],
-    'require-jsdoc': ['error', {
-      require: {
-        FunctionDeclaration: true,
-        MethodDefinition: true,
-        ClassDeclaration: false
+    'require-jsdoc': [
+      'error',
+      {
+        require: {
+          FunctionDeclaration: true,
+          MethodDefinition: true,
+          ClassDeclaration: false
+        }
       }
-    }],
-    'semi-spacing': ['error', {
-      before: false,
-      after: false
-    }],
+    ],
     'sort-vars': 'error',
-    'space-before-function-paren': ['error', 'never'],
-    'spaced-comment': ['error', 'always'],
-    'switch-colon-spacing': 'error',
-    'wrap-regex': 'error'
+    'spaced-comment': ['error', 'always']
   }
 };